Initial TceClassModel proto/structure.
Not sure if this is what I want to go with in the long run.

For the container classes, the shared interface is really nice: you can
continue asking for children lazily all the way down.

Protos for these classes are a little tricky, since you either have all
of the data in a single massive proto, or you don't have it at all.

Trying TceClassModel at the bottom of the hierarchy to see how it
feels. Might just revert it all at the end of the day if it doesn't
turn out well, but the proto semantics are nice for the heavy-parsing
situations.

Aside: login/session auth seems to potentially be broken right now.
Need to investigate.